Рефакторинг и добавление ООП.
