+++ src/main/java/InMemoryHistoryManager.java
@@ -1,0 +1,23 @@
+import java.util.ArrayList;
+import java.util.List;
+
+public class InMemoryHistoryManager implements HistoryManager {
+
+    // Храним просмотренные задачи, максимум 10
+    private static final int HISTORY_LIMIT = 10;
+    private final List<Task> history = new ArrayList<>();
+
+    @Override
+    public void add(Task task) {
+        if (task == null) return;
+        if (history.size() == HISTORY_LIMIT) {
+            history.remove(0);
+        }
+        history.add(task);
+    }
+
+    @Override
+    public List<Task> getHistory() {
+        return new ArrayList<>(history);
+    }
+}
