+++ src/main/java/InMemoryTaskManager.java
@@ -1,0 +1,215 @@
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class InMemoryTaskManager implements TaskManager {
+    private final Map<Integer, Task> tasks = new HashMap<>();
+    private final Map<Integer, Epic> epics = new HashMap<>();
+    private final Map<Integer, Subtask> subtasks = new HashMap<>();
+    private int nextId = 0;
+
+    // Подключаем менеджер истории через интерфейс
+    private final HistoryManager historyManager;
+
+    public InMemoryTaskManager() {
+        this.historyManager = Managers.getDefaultHistory();
+    }
+
+    @Override
+    public void createTask(Task task) {
+        task.setId(nextId++);
+        tasks.put(task.getId(), task);
+    }
+
+    @Override
+    public void createEpic(Epic epic) {
+        epic.setId(nextId++);
+        epics.put(epic.getId(), epic);
+    }
+
+    @Override
+    public void createSubtask(Subtask subtask) {
+        subtask.setId(nextId++);
+        subtasks.put(subtask.getId(), subtask);
+        Epic epic = epics.get(subtask.getEpicId());
+        if (epic != null) {
+            epic.addSubtaskId(subtask.getId());
+            updateEpicStatus(epic);
+        }
+    }
+
+    @Override
+    public List<Task> getAllTasks() {
+        return new ArrayList<>(tasks.values());
+    }
+
+    @Override
+    public List<Epic> getAllEpics() {
+        return new ArrayList<>(epics.values());
+    }
+
+    @Override
+    public List<Subtask> getAllSubtasks() {
+        return new ArrayList<>(subtasks.values());
+    }
+
+    @Override
+    public void removeAllTasks() {
+        tasks.clear();
+    }
+
+    @Override
+    public void removeAllEpics() {
+        epics.clear();
+        subtasks.clear();
+    }
+
+    @Override
+    public void removeAllSubtasks() {
+        for (Epic epic : epics.values()) {
+            epic.getSubtaskIds().clear();
+            updateEpicStatus(epic);
+        }
+        subtasks.clear();
+    }
+
+    @Override
+    public Task getTaskById(int id) {
+        Task task = tasks.get(id);
+        historyManager.add(task); // запись в историю
+        return task;
+    }
+
+    @Override
+    public Epic getEpicById(int id) {
+        Epic epic = epics.get(id);
+        historyManager.add(epic);
+        return epic;
+    }
+
+    @Override
+    public Subtask getSubtaskById(int id) {
+        Subtask subtask = subtasks.get(id);
+        historyManager.add(subtask);
+        return subtask;
+    }
+
+    @Override
+    public void removeTaskById(int id) {
+        tasks.remove(id);
+    }
+
+    @Override
+    public void removeEpicById(int id) {
+        Epic epic = epics.remove(id);
+        if (epic != null) {
+            for (Integer subtaskId : epic.getSubtaskIds()) {
+                subtasks.remove(subtaskId);
+            }
+        }
+    }
+
+    @Override
+    public void removeSubtaskById(int id) {
+        Subtask subtask = subtasks.remove(id);
+        if (subtask != null) {
+            Epic epic = epics.get(subtask.getEpicId());
+            if (epic != null) {
+                epic.removeSubtaskId(id);
+                updateEpicStatus(epic);
+            }
+        }
+    }
+
+    @Override
+    public void updateTask(Task updatedTask) {
+        if (tasks.containsKey(updatedTask.getId())) {
+            tasks.put(updatedTask.getId(), updatedTask);
+        }
+    }
+
+    @Override
+    public void updateEpic(Epic updatedEpic) {
+        if (epics.containsKey(updatedEpic.getId())) {
+            Epic currentEpic = epics.get(updatedEpic.getId());
+            updatedEpic.getSubtaskIds().clear();
+            updatedEpic.getSubtaskIds().addAll(currentEpic.getSubtaskIds());
+            epics.put(updatedEpic.getId(), updatedEpic);
+            updateEpicStatus(updatedEpic);
+        }
+    }
+
+    @Override
+    public void updateSubtask(Subtask updatedSubtask) {
+        if (subtasks.containsKey(updatedSubtask.getId())) {
+            Subtask oldSubtask = subtasks.get(updatedSubtask.getId());
+            int oldEpicId = oldSubtask.getEpicId();
+            int newEpicId = updatedSubtask.getEpicId();
+            if (oldEpicId != newEpicId) {
+                Epic oldEpic = epics.get(oldEpicId);
+                if (oldEpic != null) {
+                    oldEpic.removeSubtaskId(oldSubtask.getId());
+                    updateEpicStatus(oldEpic);
+                }
+                Epic newEpic = epics.get(newEpicId);
+                if (newEpic != null) {
+                    newEpic.addSubtaskId(updatedSubtask.getId());
+                }
+            }
+            subtasks.put(updatedSubtask.getId(), updatedSubtask);
+            Epic epic = epics.get(updatedSubtask.getEpicId());
+            if (epic != null) {
+                updateEpicStatus(epic);
+            }
+        }
+    }
+
+    @Override
+    public List<Subtask> getSubtasksOfEpic(int epicId) {
+        Epic epic = epics.get(epicId);
+        List<Subtask> result = new ArrayList<>();
+        if (epic == null) return result;
+        for (Integer subtaskId : epic.getSubtaskIds()) {
+            Subtask subtask = subtasks.get(subtaskId);
+            if (subtask != null) {
+                result.add(subtask);
+            }
+        }
+        return result;
+    }
+
+    @Override
+    public List<Task> getHistory() {
+        return historyManager.getHistory();
+    }
+
+    private void updateEpicStatus(Epic epic) {
+        List<Integer> subtaskIds = epic.getSubtaskIds();
+        if (subtaskIds.isEmpty()) {
+            epic.setStatus(Status.NEW);
+            return;
+        }
+        boolean allDone = true;
+        boolean allNew = true;
+        for (Integer subtaskId : subtaskIds) {
+            Subtask subtask = subtasks.get(subtaskId);
+            if (subtask == null) {
+                continue;
+            }
+            if (subtask.getStatus() != Status.DONE) {
+                allDone = false;
+            }
+            if (subtask.getStatus() != Status.NEW) {
+                allNew = false;
+            }
+        }
+        if (allDone) {
+            epic.setStatus(Status.DONE);
+        } else if (allNew) {
+            epic.setStatus(Status.NEW);
+        } else {
+            epic.setStatus(Status.IN_PROGRESS);
+        }
+    }
+}
