+++ src/main/java/Managers.java
@@ -1,0 +1,12 @@
+public class Managers {
+
+    // Возвращает реализацию менеджера задач
+    public static TaskManager getDefault() {
+        return new InMemoryTaskManager();
+    }
+
+    // Возвращает реализацию менеджера истории
+    public static HistoryManager getDefaultHistory() {
+        return new InMemoryHistoryManager();
+    }
+}
