+++ src/main/java/Task.java
@@ -1,3 +1,8 @@
+import lombok.Getter;
+import lombok.Setter;
+
+@Setter
+@Getter
 public class Task {
     private int id;             // Идентификатор задачи
     private String name;        // Название задачи
@@ -8,38 +13,6 @@ public class Task {
         this.name = name;
         this.description = description;
         this.status = Status.NEW; // По умолчанию новая задача
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Status getStatus() {
-        return status;
-    }
-
-    public void setStatus(Status status) {
-        this.status = status;
     }
 
     @Override
